Extracted class from DependencyReportTask.
